feat: update all items lists to handle crud delete actions and refresh table post-deletion
